Remove BookVM mapping in CategoryService
This commit removes the transformation of `Books` from a collection of `Book` entities to a list of `BookVM` view models in the `CategoryService.cs` file. The mapping of properties (Id, Title, Slug, ImagePath) for each `Book` object has been eliminated, indicating that the `Books` property will no longer be represented in a view model format within the `Category` entity.