Add CategoryBookViewComponent for view rendering
Introduces a new `CategoryBookViewComponent` class in the
`Refhub_Ir.Components.Handlers.Home` namespace. The class
includes a constructor that accepts an `IBookService` and
implements the `InvokeAsync` method to return a view
asynchronously based on a default view path. Necessary
using directives for `Microsoft.AspNetCore.Mvc` and
`Refhub_Ir.Service.Interface` are also added.